Nan support
Add support for finding node modules, and specifically the Nan module,
which is a common / recommended dependency for native modules that helps
standardize the native interface across different versions of node.js.
This prioritizes the use of the Nan module installed by the user, but
falls back to the one provided by this module.