Windows delayload and IOJS support
Added delay loading hook for windows to handle binary renaming on that
platform. Had to adjust the variant scripts to support populating a list
of executables regardless of which variant is selected.

Added IOJS variant to test the delayload generation. Variants now
populate variables that are conditionally populated by the NodeJS
variant if unset. This allows variants like IOJS, which is nearly
identical to NodeJS, to use this default behavior while tuning it for
the specifics of the variant.