Module integration and scripts
Lots of changes to make module building as seamless as it is in
node-gyp. Introduces a new script called 'ncmake' that automates the
creation of cmake build folders and invokes cmake with the correct
argument. Its provided as a convienence, but isn't required to build
(cmake still works on its own, there is no coupling).