Variant migration and updates
Making a few changes to support NWJS and Electron as variants for the
build system. Primarily, small flags to deal with their differences in
structure and architecture of their header sources.

Also adding long desired feature support for the 'latest' version, that
derives the NodeJS version to use for a specific variant when the
version 'latest' is used. Bit tricky for Electron which is consistently
the outlier, but found a workaround via it's github repo.

Lots of little refinements to make building and re-building more
consistent when settings changed. Generally, the policy is now that
settings are persistent once set, and are updated when set again. This
means that you have to re-run settings to get back to the defaults, or,
thanks to CMake, just delete the build directory and invoke again (via
npm install or ncmake).