Rewrite for 2.0
Complete refactor of the build system to make several improvements to
the module's operation. Primarily, the focus was to shfit back to a more
node-gyp compliant interface, to ease integration into existing
projects. This also enables compatibility with other tools that expect
node-gyp behaviour.

Condensed the entire node configuration into a single set of functions
in a single file. The module loading techinque used in version 1.0 is
error prone with specific, nuanced module hierarchies. The new method is
to copy this file directly into your project. This won't automatically
stay in sync when updating the module, but resolves this issue without
requiring a more complex user configuration to use it.

Variants have been removed, replaced with a general-purpose system that
is compatible with the Node.js release server structure. Each release
must contain a checksum file (SHASUMS256.txt) that contains hashes for
all files hosted by the server. Headers must be available in a dedicated
archive, named by the conventions on the standard server
(<name>-<version>-headers.tar.gz).