Adding update command
To make updating the module easier, adding an update command to ncmake
that copies the latest NodeJS.cmake from the module to the current
working directory. This command should be run each time the module is
updated to ensure that the latest script is in use.

A symlink would be optimal as an alternative for this, but won't work
under windows, and is poorly supported in git (AFAIK).