LINK_FLAGS hot redux
APPEND_STRING doesn't space separate arguments added to properties. Use
a prefix space so this works properly even if LINK_FLAGS already
contains flags. This makes the assumption that extra spaces in the
beginning of the property won't be an issue; this would need to be
fancier if that ever is the case.