Ncmake updates
Adding a few requested options to ncmake. Specifically, this attempts to
provide better alignment with the compiler architecture on windows,
attempting to pass MSVC an architecture parameter based on the version
of node used to execute ncmake.

Also includes the ability to specify the build output directory. Note
that this option will need to be specified with every command, and will
break the search algorithm provided with this module. This is only meant
for very specific use cases where these caveats are understood. The
default matches the old behaviour, so nothing will change if
unspecified.

Finally, adding `install` into to the set of commands. This doesn't do
what node-gyp originally did, instead opting to do the equivalent of
`cmake --build <output> --target install`. By default, this will fail as
no install definitions are created by NodeJS.cmake.