Ncmake architecture improvements
Modifying ncmake to pass the architecture flag even when the generator
is unspecified. This makes the assumption that Visual Studio is the
default generator on windows (which should always be true), but
continues to address the 32-bit default module issue.

Also, added some debug output to the utility so that I can actually see
what is going on when things are executed, and ask other people for
feedback.